# openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: info@librematch.org
    name: Libre:Match Team
    url: https://librematch.org
  description:
    "stable, reliable and feature-rich API for match information, leaderboards,\
    \ and tournament organisation for the Age of Empires multiplayer community"
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl.txt
    x-last-modified: 1663947772699
  title: Libre:Match API
  version: 1.0.0-DRAFT
  x-logo:
    url: ""
servers:
  - description: Libre:Match Server v1
    url: https://api.librematch.org/v1
    x-last-modified: 1663947810574
security:
  - BasicAuth: []
  - X-Api-Key: []
paths:
...
  /games/{game_id}/matches:
    get:
      operationId: get_match_collection_for_game
      parameters:
        - deprecated: false
          description:
            The unique identifier (ULID) we use for games that use the Relic
            Link API
          explode: false
          in: path
          name: game_id
          required: true
          schema:
            $ref: "#/components/schemas/Ulid"
          style: simple
          x-last-modified: 1663949717971
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Retrieve all matches for a specific game
      tags:
        - Match data
    summary:
      "match data spanning over the complete game (all leaderboards, basically\
      \ a log"
    x-last-modified: 1663948210036
...
  /profiles/{relic_link_id}/matches:
    get:
      operationId: get_match_collection_for_profile
      parameters:
        - deprecated: false
          description:
            The unique identifier used by the Relic Link API for a player
            on their platform
          explode: false
          in: path
          name: relic_link_id
          required: true
          schema:
            $ref: "#/components/schemas/RelicLinkId"
          style: simple
          x-last-modified: 1663949613648
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Retrieve all matches for a specific profile
      tags:
        - Match data
    summary: match data for a certain Relic Link profile
    x-last-modified: 1663948607956
...
  /leaderboards/{leaderboard_id}/matches:
    get:
      operationId: get_match_collection_for_leaderboard
      parameters:
        - deprecated: false
          description:
            The unique identifier (ULID) we use for leaderboards of the Relic
            Link API
          explode: false
          in: path
          name: leaderboard_id
          required: true
          schema:
            $ref: "#/components/schemas/Ulid"
          style: simple
          x-last-modified: 1663949692803
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Retrieve all matches on a specific leaderboard for a certain game
      tags:
        - Leaderboards
        - Match data
    summary: Collection of match data for a specific leaderboard
    x-last-modified: 1665053682780
...
  /tournaments/{tournament_id}/matches:
    get:
      operationId: get_match_collection_for_tournament
      parameters:
        - deprecated: false
          description: The unique identifier (ULID) we use for tournaments
          explode: false
          in: path
          name: tournament_id
          required: true
          schema:
            $ref: "#/components/schemas/Ulid"
          style: simple
          x-last-modified: 1663949674724
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Retrieve all match data for a specific tournament for a specific game
      tags:
        - Match data
        - Tournaments
        - Tournament Administration
    summary: game-specific tournaments
    x-last-modified: 1665054963238
...
  /teams/{team_id}/matches:
    get:
      operationId: get_match_collection_for_team
      parameters:
        - deprecated: false
          description:
            The unique identifier (ULID) we use for teams that play on games
            using the Relic Link API
          explode: false
          in: path
          name: team_id
          required: true
          schema:
            $ref: "#/components/schemas/Ulid"
          style: simple
          x-last-modified: 1665069480649
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "501":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
          description: Not Implemented
          x-last-modified: 1665058119290
      summary: Retrieve a list of matches for a specific team
      tags:
        - Match data
    summary: Collection of match information for a team
    x-last-modified: 1665139050992
...
components:
...
  schemas:
...
    RelicLinkId:
      additionalProperties: false
      properties:
        relic_link_id:
          format: int64
          maximum: 999999999
          minimum: 0
          type: integer
      required:
        - relic_link_id
      type: object
      x-last-modified: 1665056674054
    Ulid:
      additionalProperties: false
      properties:
        ulid:
          format: ulid
          maxLength: 26
          maximum: 26
          minimum: 26
          pattern: "^[0-7][0-9A-HJKMNP-TV-Z]{25}$"
          type: string
      required:
        - ulid
      type: object
      x-last-modified: 1665056674054
...
    NotImplementedError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 501
          minimum: 501
          title: code
          type: integer
      required:
        - code
        - message
      title: NotImplementedError
      type: object
      x-last-modified: 1665056674054
    TooManyRequestsError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 429
          minimum: 429
          title: code
          type: integer
      required:
        - code
        - message
      title: TooManyRequestsError
      type: object
      x-last-modified: 1665056674054
...
    NotAcceptableError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 406
          minimum: 406
          title: code
          type: integer
      required:
        - code
        - message
      title: NotAcceptableError
      type: object
      x-last-modified: 1665056674054
    NotFoundError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 404
          minimum: 404
          title: code
          type: integer
      required:
        - code
        - message
      title: NotFoundError
      type: object
      x-last-modified: 1665056674054
    ForbiddenError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 403
          minimum: 403
          title: code
          type: integer
      required:
        - code
        - message
      title: ForbiddenError
      type: object
      x-last-modified: 1665056674054
...
    UnauthorizedError:
      additionalProperties: false
      properties:
        message:
          maxLength: 32
          pattern: "^example-[0-9a-z]+$"
          title: Unauthorized access
          type: string
        code:
          maximum: 401
          minimum: 401
          title: code
          type: integer
      required:
        - code
        - message
      title: UnauthorizedError
      type: object
      x-last-modified: 1665056676293
...
  securitySchemes:
    BasicAuth:
      description: "Example: \n> Authorization: <Sample>"
      in: header
      name: BasicAuth
      type: apiKey
      x-last-modified: 1663947718456
    X-Api-Key:
      description: "Example: \n> X-Api-Key: ZGVtbzpwQDU1dzByZA=="
      in: header
      name: X-Api-Key
      type: apiKey
      x-last-modified: 1663947718456
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1665071628782